agent_app/requirements.txt
agent-framework[azure]
azure-identity
python-dotenv
